requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
idna==3.10
multidict==6.1.0
outcome==1.3.0.post0
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
retrying==1.3.4
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
twilio==9.3.2
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.13.1
